requirements.txt
django>=3.0
requests>=2.23.0
django-cors-middleware>=1.3.1
django-cors-headers>=3.2.0
django-debug-toolbar==2.0
gunicorn==19.9.0
dj-database-url==0.5.0
djangorestframework>=3.10.3
djangorestframework-simplejwt>=4.3.0
whitenoise>=3.2
psycopg2>=2.7.7
psycopg2-binary>=2.7.7
Pillow>=6.2.1
django-crispy-forms>=1.9.0
